+++ tests/test_sqlite.py
@@ -1,0 +1,74 @@
+import logging
+import sqlite3
+from argparse import Namespace
+from shutil import rmtree
+from unittest import TestCase, mock
+
+import pytest
+
+from piicatcher.explorer.sqlite import SqliteExplorer
+from tests.test_databases import CommonExplorerTestCases, CommonDataTypeTestCases, pii_data_script, char_data_types
+
+
+@pytest.fixture(scope="class")
+def temp_sqlite(request, tmpdir_factory):
+    temp_dir = tmpdir_factory.mktemp("sqlite_test")
+    sqlite_path = temp_dir.join("sqldb")
+
+    explorer = SqliteExplorer(Namespace(
+        path=sqlite_path,
+        config_file=None))
+
+    request.cls.explorer = explorer
+    request.cls.path = str(sqlite_path)
+
+    def finalizer():
+        explorer.get_connection().close()
+        rmtree(temp_dir)
+        logging.info("Deleted {}".format(str(temp_dir)))
+
+    request.addfinalizer(finalizer)
+
+
+@pytest.fixture(scope="class")
+def load_pii(request, temp_sqlite):
+    conn = sqlite3.connect(request.cls.path)
+    conn.executescript(pii_data_script)
+    conn.commit()
+    conn.close()
+
+
+@pytest.mark.usefixtures("load_pii")
+class SqliteExplorerTest(CommonExplorerTestCases.CommonExplorerTests):
+    def test_schema(self):
+        names = [sch.get_name() for sch in self.explorer.get_schemas()]
+        self.assertEqual([''], names)
+
+    def get_test_schema(self):
+        return ""
+
+
+@pytest.fixture(scope="class")
+def load_char_data(request, temp_sqlite):
+    conn = sqlite3.connect(request.cls.path)
+    conn.executescript(char_data_types)
+    conn.commit()
+    conn.close()
+
+
+@pytest.mark.usefixtures("load_char_data")
+class SqliteDataTypeTest(CommonDataTypeTestCases.CommonDataTypeTests):
+    def get_test_schema(self):
+        return ""
+
+
+class TestDispatcher(TestCase):
+    def test_sqlite_dispatch(self):
+        with mock.patch('piicatcher.explorer.sqlite.SqliteExplorer.scan', autospec=True) as mock_scan_method:
+            with mock.patch('piicatcher.explorer.sqlite.SqliteExplorer.get_tabular', autospec=True) as mock_tabular_method:
+                with mock.patch('piicatcher.explorer.explorer.tableprint', autospec=True) as MockTablePrint:
+                    SqliteExplorer.dispatch(Namespace(path='connection', list_all=None, output_format='ascii_table',
+                                                      scan_type=None, config_file=None))
+                    mock_scan_method.assert_called_once()
+                    mock_tabular_method.assert_called_once()
+                    MockTablePrint.table.assert_called_once()
